golang/22_generate_parentheses_test.go
package main

import (
	"testing"
)

func Test_generateParenthesis(t *testing.T) {
	type args struct {
		n int
	}
	tests := []struct {
		name string
		args args
		want []string
	}{
		{
			name: "1",
			args: args{
				n: 3,
			},
			want: []string{"((()))", "(()())", "(())()", "()(())", "()()()"},
		},
		{
			name: "2",
			args: args{
				n: 1,
			},
			want: []string{"()"},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := generateParenthesis(tt.args.n); !areParenthesesEqual(got, tt.want) {
				t.Errorf("generateParenthesis() = %v, want %v", got, tt.want)
			}
		})
	}
}

func areParenthesesEqual(got, target []string) bool {
	m := make(map[string]interface{}, len(got))
	for _, g := range got {
		m[g] = struct{}{}
	}
	for _, t := range target {
		if _, ok := m[t]; !ok {
			return false
		}
	}
	return true
}
